tweak: Add documentation to daemonize process
when running teddit not on a docker container.

I've generally found docker gnarly, and try not to use it. In this
case, I find pm2 to be a more convenient alternative.

It might also be interesting to have better documentation for running
this from a reverse proxy, maybe step by step, so that more people
can host these instances.